Make SLURM jobs from Maestro output (no cont yet)
Add scripts that allow the user to generate and submit a SLURM job
from the output of Maestro's "write job" functionality. Does not yet
support continuing multiphase jobs without restarting the computation
(which is going to be important).

In addition, does not yet support multi-gpu that we see on Maverick.