Update subrange code for non-zero first residues
The old subselection code assumed that the first residue in a file was
residue 1. This is not the case--for example, in 1e41, the first residue
is 89. This causes issues when interfacing with Broom's DB (e.g. when
asking for residues 98-110 in a 104 residue file where the first residue
is 89 will cause an out-of-bounds error, because the 110-th residue does
not exist.

To fix this, instead of filtering residues after the fact, we only push
Ca-atoms into the array when we know that the resnumber is inbounds of
the range.