fix: Allow backreferences between blocks
In spite of being aware that backreferences may extend between blocks,
and only backreferences past the beginning of the entire DEFLATE-encoded
stream is an error, I wrote the backreference code in a way that
referring past the beginning of a block was an error.

This commit applies the necessary restructuring to both code and tests
to fix this issue.